commandutil: introduce limit stdouterr writer and flag
This was splitted from #9978 to review separately.

Currently, our executor keeps remote action stdout/stderr inside
in-memory buffers. This approach can cause OOM-kill when the action
stdout/stderr are large.

Introduce a new flag to control the size of stdouterr of remote actions.
Also coupled with it is a new writer struct which returns a
ResourceExhaustedError when the limit is reached.

In a subsequent PR, we shall use this writer to modify different
isolation implementations to control the memory usage of Executor.